backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-telegram-bot
google-genai
python-dotenv
httpx
pytz
aiofiles
PyJWT
google-auth
google-auth-httplib2

